Rename loop_card folder for consistency
