+++ lib/music_player/looper/loop_card.dart
@@ -1,8 +1,8 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:musbx/music_player/loop_card/loop_slider.dart';
-import 'package:musbx/music_player/loop_card/looper.dart';
+import 'package:musbx/music_player/looper/loop_slider.dart';
+import 'package:musbx/music_player/looper/looper.dart';
 import 'package:musbx/music_player/music_player.dart';
 
 class LoopCard extends StatelessWidget {
