+++ lib/music_player/music_player_screen.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:musbx/music_player/demixer/demixer_card.dart';
 import 'package:musbx/music_player/equalizer/equalizer_card.dart';
-import 'package:musbx/music_player/loop_card/loop_card.dart';
+import 'package:musbx/music_player/looper/loop_card.dart';
 import 'package:musbx/music_player/music_player.dart';
 import 'package:musbx/music_player/slowdowner/slowdowner_card.dart';
 import 'package:musbx/music_player/position_card/button_panel.dart';
