+++ lib/music_player/position_card/position_slider.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:musbx/music_player/loop_card/looper.dart';
+import 'package:musbx/music_player/looper/looper.dart';
 import 'package:musbx/music_player/position_card/highlighted_section_slider_track_shape.dart';
 import 'package:musbx/music_player/music_player.dart';
